backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
motor==2.5.1
python-decouple==3.6
pydantic==1.9.1